[NEW] Rename of the folder of kafka
